front: Move Role Assigner Dropdown, remove Practice dropdown
