+++ front/src/Settings/index.js
@@ -1,19 +1,40 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Turtle from "../Turtle";
+import styled from "styled-components";
+import Dropdown from "../Dropdown";
 
 /**
  * Show the settings of all turtles.
  * Author: H.E. van der Laan
+ * Author: T.T.P. Franken
  *
  * props:
  * - turtles: an array of Turtles
  * - onTurtleEnableChange: function to call when the turtle enable button is pressed
+ * - onSettingsChange: a function which dictates what happens when a dropdown is changed.
  */
+
+const CONFIG_VALUES = [
+  "Role assigner on",
+  "Role assigner off",
+  "Pass demo",
+  "Penalty mode",
+  "Ball Handling demo"
+];
+
 const Settings = props => {
-  const { turtles } = props;
+  const { turtles, onChange } = props;
   return (
-    <div>
+    <SettingsWrapper>
+      <RefboxDropdown
+        currentValue={"Whatever"}
+        values={CONFIG_VALUES}
+        onChange={value => {
+          onChange("role", value);
+        }}
+        enabled={true}
+      />
       {turtles.map(turtle => (
         <Turtle
           key={turtle.id}
@@ -22,7 +43,7 @@ const Settings = props => {
           onChange={(changedProp, newValue) => {} /* TODO: turtle update */}
         />
       ))}
-    </div>
+    </SettingsWrapper>
   );
 };
 
@@ -36,7 +57,20 @@ Settings.propTypes = {
       role: PropTypes.string.isRequired,
       team: PropTypes.string.isRequired
     })
-  ).isRequired
+  ).isRequired,
+  onChange: PropTypes.func.isRequired
 };
+
+const RefboxDropdown = styled(Dropdown)`
+  width: 75%;
+  margin-left: auto;
+  margin-right: auto;
+  height: 4rem;
+`;
+
+const SettingsWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
 
 export default Settings;
